+++ src/main/java/ru/coffeecoders/questbot/commands/admins/keybords/keyboards/AdminsCommandActions.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Component;
 @Component
 public class AdminsCommandActions {
 
-
-
+// 2.TODO логика открытия и закрытия клавиатур
+    //1. TODO new ReplyKeyboardRemove()
