+++ src/main/java/ru/coffeecoders/questbot/commands/admins/keybords/keyboards/KeyboardSender.java
@@ -2,8 +2,6 @@ package ru.coffeecoders.questbot.commands.admins.keybords.keyboards;
 
 
 import com.pengrad.telegrambot.model.Update;
-import com.pengrad.telegrambot.model.request.ReplyKeyboardMarkup;
-import com.pengrad.telegrambot.request.SendMessage;
 import org.springframework.stereotype.Component;
 import ru.coffeecoders.questbot.commands.admins.keybords.keyboards.creators.Keyboard;
 
@@ -12,7 +10,7 @@ import ru.coffeecoders.questbot.commands.admins.keybords.keyboards.creators.Keyb
 
 public class KeyboardSender implements Keyboard {
 
-    //TODO ещё думаю как реализовать
+    //TODO сборка для ответа в messageSender
 
     public void sendKeyboard(Keyboard keyboard, Update update) {
         messageSender.sendMessage(keyboard, update);
