+++ src/main/java/ru/coffeecoders/questbot/commands/admins/keybords/keyboards/creators/AllQuestionsKeyboardCreator.java
@@ -4,32 +4,23 @@ import com.pengrad.telegrambot.model.request.InlineKeyboardButton;
 import com.pengrad.telegrambot.model.request.InlineKeyboardMarkup;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import ru.coffeecoders.questbot.entities.Question;
-import ru.coffeecoders.questbot.repositories.QuestionRepository;
-
-import java.util.ArrayList;
-
-import java.util.List;
 
 @Component
-public class AllQuestionsCreator extends InlineKeyboardMarkup implements Keyboard{
+public class AllQuestionsKeyboardCreator extends InlineKeyboardMarkup implements Keyboard{
     @Value("${keyboard.allQuestions.edit}")
     private static String edits;
 
    @Value("${keyboard.allQuestions.toGroup}")
     private static String deletes;
+    private AllQuestionsKeyboardCreator allQuestionsKeyboardCreator;
 
-    private List<Question> allQuestions = new ArrayList<>();
-    private List<List<Question>> paginatedQuestions = new ArrayList<>();
-    private int currentPage = 0;
-    private AllQuestionsCreator allQuestionsCreator;
 
-    private AllQuestionsCreator(AllQuestionsCreator allQuestionsCreator) {
-        this.allQuestionsCreator = allQuestionsCreator;
+    private AllQuestionsKeyboardCreator(AllQuestionsKeyboardCreator allQuestionsKeyboardCreator) {
+        this.allQuestionsKeyboardCreator = allQuestionsKeyboardCreator;
     }
 
-    public AllQuestionsCreator getAllQuestionsCreator() {
-        return allQuestionsCreator;
+    public AllQuestionsKeyboardCreator getAllQuestionsCreator() {
+        return allQuestionsKeyboardCreator;
     }
 
     public static InlineKeyboardMarkup inlineKeyboardCreate() {
@@ -64,24 +55,6 @@ public class AllQuestionsCreator extends InlineKeyboardMarkup implements Keyboar
     }
 
     //TODO нужно перененести эту логику, попробую сделать
-    public void resizeQuestionsList() {
-
-        paginatedQuestions = QuestionService.allQuestionsBySix(allQuestions);
-    }
-
-    public List<List<Question>>allQuestionsBySix(List<Question> allQuestion) {
-        allQuestions = QuestionRepository.getAllQuestion();
-        List<List<Question>> paginatedQuestions = new ArrayList<>();
-        int page = 0;
-        while (page * 6 < allQuestions.size()) {
-            int fromIndex = page * 6;
-            int toIndex = Math.min((page + 1) * 6, allQuestions.size());
-            paginatedQuestions.add(allQuestions.subList(fromIndex, toIndex));
-            page++;
-        }
-
-        return paginatedQuestions;
-    }
 
 
 }
